requirements.txt
feedparser==6.0.11
discord-webhook==1.3.1
python-dotenv==1.0.0
requests==2.31.0